requirements.txt
pillow
pypdf
ocrmypdf
